+++ frontends/php/include/discovery.inc.php
@@ -1,7 +1,7 @@
 <?php
 /*
 ** ZABBIX
-** Copyright (C) 2000-2010 SIA Zabbix
+** Copyright (C) 2000-2005 SIA Zabbix
 **
 ** This program is free software; you can redistribute it and/or modify
 ** it under the terms of the GNU General Public License as published by
@@ -20,9 +20,10 @@
 ?>
 <?php
 
-	require_once('include/perm.inc.php');
+require_once('include/perm.inc.php');
 
-
+?>
+<?php
 	function check_right_on_discovery($permission){
 		global $USER_DETAILS;
 
@@ -34,6 +35,8 @@
 	}
 
 	function svc_default_port($type_int){
+		$port = '0';
+
 		switch($type_int){
 			case SVC_SSH:		$port = '22';		break;
 			case SVC_LDAP:		$port = '389';		break;
@@ -49,31 +52,31 @@
 			case SVC_SNMPv2:	$port = '161';		break;
 			case SVC_SNMPv3:	$port = '161';		break;
 			case SVC_ICMPPING:	$port = '0';		break;
-			default: 			$port = '0';		break;
 		}
 
-		return $port;
+	return $port;
 	}
 
 	function discovery_check_type2str($type_int){
-		$types = array(
-			SVC_SSH => S_SSH,
-			SVC_LDAP => S_LDAP,
-			SVC_SMTP => S_SMTP,
-			SVC_FTP => S_FTP,
-			SVC_HTTP => S_HTTP,
-			SVC_POP => S_POP,
-			SVC_NNTP => S_NNTP,
-			SVC_IMAP => S_IMAP,
-			SVC_TCP => S_TCP,
-			SVC_AGENT => S_ZABBIX_AGENT,
-			SVC_SNMPv1 => S_SNMPV1_AGENT,
-			SVC_SNMPv2 => S_SNMPV2_AGENT,
-			SVC_SNMPv3 => S_SNMPV3_AGENT,
-			SVC_ICMPPING => S_ICMPPING,
-		);
-		
-		return isset($types[$type_int]) ? $types[$type_int] : S_UNKNOWN;
+		$str_type[SVC_SSH]	= S_SSH;
+		$str_type[SVC_LDAP]	= S_LDAP;
+		$str_type[SVC_SMTP]	= S_SMTP;
+		$str_type[SVC_FTP]	= S_FTP;
+		$str_type[SVC_HTTP]	= S_HTTP;
+		$str_type[SVC_POP]	= S_POP;
+		$str_type[SVC_NNTP]	= S_NNTP;
+		$str_type[SVC_IMAP]	= S_IMAP;
+		$str_type[SVC_TCP]	= S_TCP;
+		$str_type[SVC_AGENT]	= S_ZABBIX_AGENT;
+		$str_type[SVC_SNMPv1]	= S_SNMPV1_AGENT;
+		$str_type[SVC_SNMPv2]	= S_SNMPV2_AGENT;
+		$str_type[SVC_SNMPv3]	= S_SNMPV3_AGENT;
+		$str_type[SVC_ICMPPING]	= S_ICMPPING;
+
+		if(isset($str_type[$type_int]))
+			return $str_type[$type_int];
+
+		return S_UNKNOWN;
 	}
 
 	function discovery_check2str($type, $snmp_community, $key_, $ports){
@@ -104,29 +107,35 @@
 	function discovery_status2str($status_int){
 		switch($status_int){
 			case DRULE_STATUS_ACTIVE:	$status = S_ACTIVE;		break;
-			case DRULE_STATUS_DISABLED:	$status = S_DISABLED;	break;
-			default:					$status = S_UNKNOWN;	break;
+			case DRULE_STATUS_DISABLED:	$status = S_DISABLED;		break;
+			default:
+				$status = S_UNKNOWN;		break;
 		}
-		return $status;
+
+	return $status;
 	}
 
 	function discovery_status2style($status){
 		switch($status){
-			case DRULE_STATUS_ACTIVE: $status = 'off'; break;
-			case DRULE_STATUS_DISABLED: $status = 'on'; break;
-			default: $status = 'unknown'; break;
+			case DRULE_STATUS_ACTIVE:	$status = 'off';	break;
+			case DRULE_STATUS_DISABLED:	$status = 'on';		break;
+			default:
+				$status = 'unknown';	break;
 		}
-		return $status;
+
+	return $status;
 	}
 
 	function discovery_object_status2str($status){
-		$statuses = array(
-			DOBJECT_STATUS_UP => S_UP,
-			DOBJECT_STATUS_DOWN => S_DOWN,
-			DOBJECT_STATUS_DISCOVER => S_DISCOVERED,
-			DOBJECT_STATUS_LOST => S_LOST,
-		);
-		return isset($statuses[$status]) ? $statuses[$status] : S_UNKNOWN;
+		$str_stat[DOBJECT_STATUS_UP] = S_UP;
+		$str_stat[DOBJECT_STATUS_DOWN] = S_DOWN;
+		$str_stat[DOBJECT_STATUS_DISCOVER] = S_DISCOVERED;
+		$str_stat[DOBJECT_STATUS_LOST] = S_LOST;
+
+		if(isset($str_stat[$status]))
+			return $str_stat[$status];
+
+		return S_UNKNOWN;
 	}
 
 	function get_discovery_rule_by_druleid($druleid){
@@ -138,14 +147,14 @@
 	}
 
 	function add_discovery_check($druleid, $type, $ports, $key, $snmp_community,
-			$snmpv3_securityname, $snmpv3_securitylevel, $snmpv3_authpassphrase, $snmpv3_privpassphrase, $uniq=0)
+			$snmpv3_securityname, $snmpv3_securitylevel, $snmpv3_authpassphrase, $snmpv3_privpassphrase)
 	{
 		$dcheckid = get_dbid('dchecks', 'dcheckid');
 		$result = DBexecute('insert into dchecks (dcheckid,druleid,type,ports,key_,snmp_community'.
-				',snmpv3_securityname,snmpv3_securitylevel,snmpv3_authpassphrase,snmpv3_privpassphrase,uniq) '.
+				',snmpv3_securityname,snmpv3_securitylevel,snmpv3_authpassphrase,snmpv3_privpassphrase) '.
 				' values ('.$dcheckid.','.$druleid.','.$type.','.zbx_dbstr($ports).','.
 				zbx_dbstr($key).','.zbx_dbstr($snmp_community).','.zbx_dbstr($snmpv3_securityname).','.
-				$snmpv3_securitylevel.','.zbx_dbstr($snmpv3_authpassphrase).','.zbx_dbstr($snmpv3_privpassphrase).','.$uniq.')');
+				$snmpv3_securitylevel.','.zbx_dbstr($snmpv3_authpassphrase).','.zbx_dbstr($snmpv3_privpassphrase).')');
 
 		if(!$result)
 			return $result;
@@ -161,16 +170,21 @@
 
 		$druleid = get_dbid('drules', 'druleid');
 		$result = DBexecute('insert into drules (druleid,proxy_hostid,name,iprange,delay,status) '.
-			' values ('.$druleid.','.zero2null($proxy_hostid).','.zbx_dbstr($name).','.zbx_dbstr($iprange).','.$delay.','.$status.')');
+			' values ('.$druleid.','.$proxy_hostid.','.zbx_dbstr($name).','.zbx_dbstr($iprange).','.$delay.','.$status.')');
 
 		if($result && isset($dchecks)){
 			$unique_dcheckid = 0;
 			foreach($dchecks as $id => $data){
 				$data['dcheckid'] = add_discovery_check($druleid, $data['type'], $data['ports'], $data['key'],
 						$data['snmp_community'], $data['snmpv3_securityname'], $data['snmpv3_securitylevel'],
-						$data['snmpv3_authpassphrase'], $data['snmpv3_privpassphrase'], 
-						($uniqueness_criteria == $id) ? 1 : 0);
+						$data['snmpv3_authpassphrase'], $data['snmpv3_privpassphrase']);
+				if ($uniqueness_criteria == $id && $data['dcheckid'])
+					$unique_dcheckid = $data['dcheckid'];
 			}
+			DBexecute('UPDATE drules'.
+					' SET unique_dcheckid='.$unique_dcheckid.
+					' WHERE druleid='.$druleid);
+
 			$result = $druleid;
 		}
 
@@ -178,12 +192,13 @@
 	}
 
 	function update_discovery_rule($druleid, $proxy_hostid, $name, $iprange, $delay, $status, $dchecks,	$uniqueness_criteria, $dchecks_deleted){
-		if(!validate_ip_range($iprange)){
+		if( !validate_ip_range($iprange) ){
 			error(S_INCORRECT_IP_RANGE);
 			return false;
+
 		}
 
-		$result = DBexecute('update drules set proxy_hostid='.zero2null($proxy_hostid).',name='.zbx_dbstr($name).',iprange='.zbx_dbstr($iprange).','.
+		$result = DBexecute('update drules set proxy_hostid='.$proxy_hostid.',name='.zbx_dbstr($name).',iprange='.zbx_dbstr($iprange).','.
 			'delay='.$delay.',status='.$status.' where druleid='.$druleid);
 
 		if($result && isset($dchecks)){
@@ -194,17 +209,13 @@
 							$data['snmp_community'], $data['snmpv3_securityname'], $data['snmpv3_securitylevel'],
 							$data['snmpv3_authpassphrase'], $data['snmpv3_privpassphrase']);
 				}
-				if($uniqueness_criteria == $id && $data['dcheckid'])
+				if ($uniqueness_criteria == $id && $data['dcheckid'])
 					$unique_dcheckid = $data['dcheckid'];
 			}
-			
-			$sql = 'UPDATE dchecks SET uniq=0 WHERE druleid='.$druleid;
-			DBexecute($sql);
 
-			if($unique_dcheckid){
-				$sql = 'UPDATE dchecks SET uniq=1 WHERE dcheckid='.$unique_dcheckid;
-				DBexecute($sql);
-			}
+			DBexecute('UPDATE drules'.
+					' SET unique_dcheckid='.$unique_dcheckid.
+					' WHERE druleid='.$druleid);
 		}
 
 		if($result && isset($dchecks_deleted) && !empty($dchecks_deleted))
@@ -237,10 +248,14 @@
 					' AND '.DBcondition('value', $dcheckids, false, true));	// FIXED[POSIBLE value type violation]!!!
 		}
 
+		DBexecute('DELETE FROM dservices WHERE '.DBcondition('dcheckid', $dcheckids));
+
 		DBexecute('DELETE FROM dchecks WHERE '.DBcondition('dcheckid', $dcheckids));
 	}
 
 	function delete_discovery_rule($druleid){
+		$result = true;
+
 		$actionids = array();
 // conditions
 		$sql = 'SELECT DISTINCT actionid '.
@@ -264,7 +279,23 @@
 					" AND value='$druleid'");
 		}
 
-		$result = DBexecute('delete from drules where druleid='.$druleid);
+		if($result){
+			$db_dhosts = DBselect('select dhostid from dhosts'.
+					' where druleid='.$druleid.' and '.DBin_node('dhostid'));
+
+			while ($result && ($db_dhost = DBfetch($db_dhosts)))
+				$result = DBexecute('delete from dservices where'.
+						' dhostid='.$db_dhost['dhostid']);
+		}
+
+		if ($result)
+			$result = DBexecute('delete from dhosts where druleid='.$druleid);
+
+		if ($result)
+			$result = DBexecute('delete from dchecks where druleid='.$druleid);
+
+		if ($result)
+			$result = DBexecute('delete from drules where druleid='.$druleid);
 
 		return $result;
 	}
